refactor: 배경에서 안내문구를 보여주던 AdvisorFrame을 컴포넌트로 분리
